requirements.txt
flask
python-dotenv
langchain
langchain-openai
faiss-cpu
gunicorn